Reworked the rebuildall scripts and added new build instructions.
The rebuildall script allow you to build only part of Cerberus if you need to without having to do any manual swapping and changing. Transcc building will make a back up just in case the build breaks so it can be restored.
They can also build distribution packages if git is installed in the paths.