Update to transcc for Linux GCC 6+ issue and PIE's.
Some file explorers on newer version of Linux have trouble detecting and running Position Independent Executable's (PIE).
To work round  this problem on Linux, versions of GCC greater or equal to version 6 are detected and the option -no-pie is passed to the back-end compiler.
This hopefully will be a temporary  measure when either they fix the problem with the file explorers, or a easy to implement deployment method is worked out.